Update create package script

+++ create-package.sh
@@ -7,7 +7,18 @@
 root_dir="$(cd $(dirname $0) && pwd -P)"
 plugin_id="plugin.video.kodiconnect"
 version="$(cat "${root_dir}/version")"
-package_filepath="${root_dir}/${plugin_id}-${version}.zip"
+python_addon_version="${PYTHON_ADDON_VERSION}"
+
+if [ -z "${python_addon_version}" ]; then
+  echo "Missing PYTHON_ADDON_VERSION variable"
+  exit 1
+fi
+
+package_dir="${root_dir}/packages/${python_addon_version}"
+
+mkdir -p "${package_dir}"
+
+package_filepath="${package_dir}/${plugin_id}-${version}.zip"
 
 rm -fv "${package_filepath}"
 
@@ -30,11 +41,16 @@ sed_replace_cmd="s/xversionx/${version}/g"
 sed -i "${sed_replace_cmd}" "${install_dir}/addon.xml"
 find "${install_dir}/resources/language" -name '*.po' -exec sed -i "${sed_replace_cmd}" "{}" \;
 
+sed_replace_cmd="s/xpythonversionx/${python_addon_version}/g"
+sed -i "${sed_replace_cmd}" "${install_dir}/addon.xml"
+
 pushd "${tmpdir}"
 
 zip -r "${package_filepath}" .
 
 popd
+
+cat "${tmpdir}/plugin.video.kodiconnect/addon.xml"
 
 rm -r "${tmpdir}"
